fusebox: Add FUSE client frontend
FuseBox needs user-space (client) code to 1) create and start a FUSE
user-space session, and 2) read and process Kernel FUSE requests for
the session.

A FuseMount object is provided to the FuseFrontend class, containing
the active |mountpoint| name and Kernel FUSE channel |chan|, and the
|args| needed to create the session.

FuseFrontend::CreateFuseSession() creates the FUSE session, and then
StartFuseSession() can be used to start and run the session.

The session is connected to Kernel FUSE over the provided |chan| and
reads Kernel FUSE requests from the |chan| file descriptor and sends
them to libFUSE, where they are processed into FUSE operations. This
is handled by FuseFrontend::OnFuseChannelReadable().

Note an EINTR error while reading the channel can be ignored: kernel
FUSE will notice and re-send requests in this case (request delivery
is reliable in FUSE).

Kernel FUSE may close the session: due to a umount(8) which unmounts
the mountpoint or by sending an error (negative read on |chan|). The
class owner is told with the stop callback, and should tear-down the
FUSE session.

Session tear-down and clean-up: class owner deletes the FuseFrontend
class to invoke its destructor.

BUG=chromium:1240983
TEST=build_packages --board=$BOARD

Change-Id: I249d078bfb099ee8a561fbde893a638941b80332
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3103050
Reviewed-by: Peter Marshall <petermarshall@chromium.org>
Reviewed-by: François Degros <fdegros@chromium.org>
Tested-by: Noel Gordon <noel@chromium.org>
Commit-Queue: Noel Gordon <noel@chromium.org>