cryptohome: Add unit test coverage for UserDataAuth::DoMount
This CL adds unit test coverage for the use case when cryptohome is
doing Public Mount, specifically when a new user is created in public
mount and when there is an error in getting a mount key.

This CL also adds support to check for CreateRequest's definition is
defined as the code expects, specifically to ensure that two keys are
not passed in CreateRequest.

BUG=b:195697285
TEST=unittest passes

Change-Id: I2c661759952e8372a64b12dae8b780977b09076f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3104209
Tested-by: Hardik Goyal <hardikgoyal@chromium.org>
Auto-Submit: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: Leo Lai <cylai@google.com>
Reviewed-by: Leo Lai <cylai@google.com>