cryptohome: [USS] unit test bad fields while load
Add test coverage for failure code paths in the UserSecretStash
parsing code, specifically to the scenarios when some of the flatbuffer
fields are missing or have bad values. Tests cover both the container
and the encrypted payload.

In order to simplify testing, we're enabling the FlatBuffer's Object
based API (only in test builds).

BUG=b:194953754
TEST=cros_workon_make --board=$BOARD --test cryptohome

Change-Id: I2bc6df4a6dbcf907c9cb0e4004695895fafcf949
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3105245
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>