lorgnette: Add a script to test scan combinations
This uses lorgnette_cli to try all listed combinations of source,
resolution, and color.  It does minimal verification on the output
files; human review after scanning it needed to make sure that output is
reasonable and any failures were expected. Does not cover things like
detecting the supported features or the valid scan areas.  This is meant
to be used for WWCB testing until we have automated this into the go
binaries.

BUG=b:201081168
TEST=Run on dut with scanner attached

Change-Id: I3f4a40f3903c6a46c1381cdbb869da93c4ccd985
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3182688
Reviewed-by: Paul Moy <pmoy@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>