vm_tools: cicerone: Make cicerone run as init
Currently cicerone runs as a child process inside a PID namespace and
has two minijail processes that stick around for it's lifetime. This
is a waste of memory, but also denies cicerone the ability to shutdown
cleanly in response to SIGTERM.

Remount /proc explicitly, since -I disabled -r.

Testing this change revealed a bug in how cicerone interacts with
libgrpc. grpc::Server::SetGlobalCallbacks assigns its argument to a
global std::shared_ptr which gets destroyed in static destruction. We
were passing it a pointer from a base::NoDestructor which can't be
deallocated like that, now we pass it the output of new.

BUG=none
TEST=Manually tested

Change-Id: I0af3796b44a6e4c3f0639f3c41259deb8f378ec5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3106327
Tested-by: Fergus Dall <sidereal@google.com>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Commit-Queue: Fergus Dall <sidereal@google.com>