camera: Only use IP HAL on devices that have the needed HW
We use the same system image for both boards that support IP peripherals
and boards that don't. That means we need to determine at run time if
the device has the hardware needed to support IP peripherals and then
either try (or not try) to connect to the IP peripheral service.

BUG=b:194120558
TEST=IP camera still works on Endeavour

Change-Id: I4027269095da349ed4d237713d76fe6c9ae97795
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3108105
Reviewed-by: Jeff Chase <jnchase@google.com>
Tested-by: Jeff Chase <jnchase@google.com>
Commit-Queue: Pablo Ceballos <pceballos@chromium.org>