crash: Put reboots due to NoC errors into different crash bucket
Sometimes when other CPUs in Qualcomm SoCs hang the bus, or devices
sitting on the bus stop working properly, the whole system hangs and the
watchdog reboots. Technically this is a watchdog reboot, but it's also
an error with the bus. Let's look for the signs that the bus, i.e. NoC
or Network On Chip, has hung by parsing the bios logs from the last boot
and make a signature out of it. We use the first NoC name to show up and
then hash together all the names and register contents to try to make
something short and unique.

BUG=b:163721897
TEST=cros_workon_make --test --board=amd64-generic crash-reporter

Change-Id: Ida5bf3c9b6a796d25c76fd34b7711836e60f6573
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3108247
Tested-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>