hermes: Open a new device before every eSIM operation
Previously, we would restart the initialization process before every
eSIM operation. We only need to reopen the device. We do not need to
reacquire the EID and IMEI.

This CL also downgrades an error that occurs because hermes starts
before the modem appears on /dev/ .

BUG=none
TEST=modem esim enable <iccid>

Change-Id: I22f1de2a99586fc29cd25b736ecf25f371e6e673
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3108138
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Pavan Holla <pholla@google.com>
Commit-Queue: Pavan Holla <pholla@google.com>