attestation: handle "}" properly in proto_print.py
The proto_print.py failed to handle "}" if it is not the only symbol
in the line. In addition, the clang-format prefer "... {}" in one line,
instead of adding an additional new line between "{}". Thus, the proto
files are not formatted by clang-format now.

This CL would fix this by enabling parser to handle the "{}", and
clang-format the proto files.

BUG=b:197479401
TEST=regenerate print_tpm_manager_proto.* by proto_print.py

Change-Id: I1968170dbfe8f64f2c56b0920a797af64c82e572
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3113403
Tested-by: Ching-Kang Yen <chingkang@chromium.org>
Commit-Queue: Ching-Kang Yen <chingkang@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>
Reviewed-by: Yi Chou <yich@google.com>