shill: wifi: refactor SignalLevel() for better unit test
Currently, SignalLevel() in wifi service is not stubbed but
the method NotifyCurrentEndpoint(which updates the current_endpoint_
, which is consumed by SignalLevel) is stubbed. So, in wifi unit
tests, the returned value of SignalLevel is always uint_16<min>(
because current_endpoint_ is always null) even when the test configures
the service to have special signal level values, and the unit tests,
DisconnectPendingServiceWith/WithOutOfRange and
TimeoutPendingServiceWithEndpoints luckily pass.

As a solution, stub SignalLevel and mock it return specific values on
which wifi tests are depending. Additionally, to maintain consistency
replace virtual with mockable for functions which are mocked for unit tests.

BUG=b:194945109
TEST=Unit test passes.

Change-Id: Icc88689308e87df8307a4380684b2f40040b2c21
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3115906
Tested-by: Abhishek Kumar <kuabhs@chromium.org>
Commit-Queue: Abhishek Kumar <kuabhs@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>