debugd: Update packet_capture command's usage
Packet capture can take --max-size as option to stop the packet capture
process when the output file reaches the given size. The option takes
input in MiB. Update the usage text in capture_utility script to have
this information in order to avoid confusion. The bug fix is aimed to
be merged to M94 with the feature.

BUG=b:197545916
TEST=Run "packet_capture" in Crosh. The output message should tell to
give --max-size input in MiBs

Change-Id: I4db9afacd7c32ef80c0d93550149a88aaf90a454
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3114659
Tested-by: Irem Uğuz <iremuguz@google.com>
Commit-Queue: Irem Uğuz <iremuguz@google.com>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>