power: cpufreq: Run even if UI doesn't come up
During bringup, my board doesn't yet boot to the UI. This meant that
the cpufreq governor was left at the default "performance"
governor. That's probably not ideal. Let's still set the cpufreq
governor even if the UI doesn't come up by using "failsafe" as talked
about in the system-services.conf file. That basically means that if
the UI doesn't come up in 30 seconds we'll still transition to the
normal governor.

This patch will have a side effect of switching us off the boot
governor slightly earlier than we did before in the _normal_ case
(when UI comes up). Specifically this is because we used to start on
"started system-services". That means we'd run after all the other
people marked to start on "starting system-services". Now we use
failsafe and failsafe starts on "starting system-services", so we'll
run in parallel to others who "start on starting
system-services". This should be OK because everything starting based
on "system-services" is supposed to a little less critical. Also,
presumably the normal cpufreq governor really shouldn't be _that_
bad. In any case, we'll still start on "started failsafe" instead of
"starting failsafe" so that anyone who really wants to run before us
could start on "starting failsafe". Looking at existing upstart
scripts that will at least make sure we start _after_ udev-trigger.

BUG=None
TEST=See that the governor changes even if UI never comes up

Change-Id: I803014b0d334f9b9c4c903ee00225c757246d5ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3116569
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>