cryptohome: [USS] Implement USS main key wrapping
Add implementation of the USS main key wrapping/unwrapping
(encryption/decryption), and a basic unit test. Advanced tests of
other failure scenarios will be done in follow-ups.

For the context, the wrapping key used for the encryption/decryption
will be coming out of the user authentication logic. In the 1FA world,
the plan is that the wrapping key will be the per-credential secret
returned by the corresponding auth block, and the wrapping ID will be
the label of that credential.

BUG=b:195776293
TEST=cros_workon_make --board=$BOARD --test cryptohome

Change-Id: Iaf563aa22dfcb8edc1616df0a1186cd0f53d9eb9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3151846
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>