hps: Add initial HpsDaemon test.
We will need to introduce mode complex logic to hpsd to support polling
of hps, debounding of the results and signaling on dbus when certain
states change. To facilitiate this we need to make it possible to unit
test these new features by mocking out the DBus layers.

This CL introduces the Mock DBus, and sets up the objects we'll need
later for more complext tests.

As an added bonus these tests uncovered a bug where HPS was returning a
bool for Enable/Disable, but HpsDaemon was treating it as an int and
hence the failure paths were not working.

Next step is to make is possible to mock HPS, so we can write more
complext tests at this level without having to mock all of the signals
at the device level.

BUG=b:197028134
TEST="FEATURE=test emerge-volteer hpsd"

Cq-Depend: chromium:3122065
Change-Id: I3cb4c8f2d987db92e8e71481d6bd24815e0c25ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3121770
Reviewed-by: Andrew McRae <amcrae@google.com>
Tested-by: Stuart Langley <slangley@chromium.org>
Commit-Queue: Stuart Langley <slangley@chromium.org>