shill: wifi: Fix out-of-range heuristic on connect timeout
The out-of-range heuristic in wifi.cc works by comparing the
AP's signal level to a threshold (currently set to -75 dBm),
and concluding that an unknown failure is caused by range if
the signal is below that threshold.

This signal level comes from WiFiService::SignalLevel() at
several points in the code. Unfortunately, this method only
actually worked when the service was currently connected,
and returned -32768 dBm the rest of the time.

This CL modifies WiFiService::SignalLevel() to return its
most recent raw_signal_strength_ estimate whenever we know
of at least one endpoint.

In the case where we're currently connected (and thus
current_endpoint_ is non-NULL) this should be a no-op because
UpdateFromEndpoints already checks current_endpoint_ when
deciding what endpoint should be 'representative' of the
service, prior to updating raw_signal_strength_.

When there are no endpoints, we continues to return -32768
and thus report an out-of-range failure if the user attempts
to connect to an AP for which we haven't seen any beacons.
Since the AP probably exists somewhere in the world this is
arguably correct behavior.

BUG=b:182090525
TEST=Verified manually on a Helios DUT (see bug for details)
TEST=unit tests

Change-Id: I2dfa874563927e797152a7c603be84165b0bbca8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/2758507
Tested-by: Jintao Lin <jintaolin@chromium.org>
Commit-Queue: Jintao Lin <jintaolin@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>