shill: alternate hidden scans when limit is 1
It was reported that the lenovo x131e with the BCM43228 wireless card
was unable to connect to hidden networks. This was because the wl driver
for devices such as this is only able to support 1 SSID per scan.

In order to fix this issue, alternate between scanning a hidden network
and doing the broadcast probe when MaxScanSSID == 1.

This change is necessary for reven builds.

BUG=b:179823373
TEST=cros_workon_make --board=reven shill --test, and this change was
also tested on CloudReady 1.0 on the lenovo x131e by first testing it
without the change to confirm that this fix is still needed, and then
testing it with the change to ensure it works as expected.

I also measured latency on the lenovo x131e. Because there is a
hidden ssid on the hidden network list, and maxScanSSID = 1 for that
chipset, that seems to add a latency of between 2s and 5s for regular
networks. Latency for the hidden network was somewhat higher, ranging
between 2s-15s. This is an expected casualty of this chipset's limited
scanning capability.

Change-Id: Iddc2bfc64b0df0685345e28bd520ade4f3f98c40
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3126325
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>
Reviewed-by: Alain Michaud <alainm@chromium.org>
Tested-by: Esther Shimanovich <eshimanovich@google.com>
Commit-Queue: Esther Shimanovich <eshimanovich@google.com>