hps: Create pure virtual HPS interface to enable testing.
As more complex functionality is added to HpsDaemon we would like to be
able to use a MockHps to make it easier to test. This cl creates a HPS
interface and a new class HPS_impl that is an implementation of that
interface - renaming the previous concrete implementation of HPS.

We also:
- Create a MockHps in hps_daemon_test and use that to add new positive
test cases.
- Where there were mixed c & c++ style comments move to c++ style.
- Instantiate HPS_Impl where previously HPS was being instantiated.
- Fix some headers that were being transitively included.

This introduces no change in functionality.

BUG=b:197028134
TEST=FEATURES=test emerge-volteer hpsd hps-tool

Change-Id: I14c28a69570846519d659310380480580fc2d2a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3126416
Tested-by: Stuart Langley <slangley@chromium.org>
Auto-Submit: Stuart Langley <slangley@chromium.org>
Reviewed-by: Andrew McRae <amcrae@google.com>
Commit-Queue: Stuart Langley <slangley@chromium.org>