hps: Add a repeating time to poll hps.
Introduce a RepeatingTimer to poll hps periodically after a feature has
been enabled, in preparation for being able to filer/debounce the
results.

We need to decouple the HpsAdaptor from the DBusDaemon so that we can
mock out the task environment so we can control time, to do this follow
the pattern described at go/pzeyt.

For the timer, we use a bitfield to know which features have been
enabled, and then only poll those features from hps.

Add basic tests to validate that this is all working.

BUG=b:198682678
TEST=FEATURES=test emerge hpsd

Change-Id: I6b039e8b7db43cebc97d8b7aa3dfd2784c3ceecf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3131541
Reviewed-by: Andrew McRae <amcrae@google.com>
Commit-Queue: Stuart Langley <slangley@chromium.org>
Tested-by: Stuart Langley <slangley@chromium.org>