cryptohome: Use new protocol in CryptohomeRecoveryAuthBlock
This CL updates the recovery auth block to using the new version
of the recovery protocol. For the background, the new protocol
uses epoch beacons and additional layers of encryption for extra
security. The RecoveryCrypto class currently exposes an
implementation of both versions, but once all callsites are
migrated to the new version the old one will be deleted.

BUG=b:194678588
TEST=cros_run_unit_tests --board=${BOARD} --packages cryptohome

Change-Id: Ice29c662182f0906e26ec5b250e32135b3e30f57
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3087447
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Greg Kerr <kerrnel@chromium.org>
Commit-Queue: Anastasiia N <anastasiian@chromium.org>