crash: Don't use hold_lock_file for death test.
In the linked bug, the test fails because hold_lock_file doesn't sleep
for long enough and the fork takes too long.

Instead of using hold_lock_file as a separate process, directly lock the
file in the parent process. This relies on the fact that the
EXPECT_EXIT() macro will start a subprocess that tries to lock the
file, and file locks are not shared between processes and subprocesses.

BUG=b:197518716
TEST=cros_workon_make --board=jacuzzi --test crash-reporter

Change-Id: I012694fbcb2a31fadebc9058a867086fe2526bf8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3130260
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>