Reland "shill: vpn: wg: Use memfd_create() to create config file"
This is a reland of 0f352df99a3226d9320ea63ecd47f08c7c80e5bb

The original patch uses base::PathExists() to check if the config file
is deleted. It seems that the file may not be deleted immediately after
the fd is closed in the driver in some cases, and thus this check incurs
some flakiness. This patch changes to check if |config_fd_| is valid
instead.

Original change's description:
> shill: vpn: wg: Use memfd_create() to create config file
>
> See comments in vpn_util.h and the bug page for details.
>
> BUG=b:198404085
> TEST=unit_tests
> TEST=tast run $DUT network.VPNConnect.wireguard*
>
> Change-Id: I63e5d236b6a64154bc63ce7433b3462e0c5e047d
> Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3135735
> Tested-by: Jie Jiang <jiejiang@chromium.org>
> Commit-Queue: Jie Jiang <jiejiang@chromium.org>
> Reviewed-by: Taoyu Li <taoyl@chromium.org>
> Reviewed-by: Hugo Benichi <hugobenichi@google.com>

BUG=b:198404085
TEST=unit_tests
TEST=tast run $DUT network.VPNConnect.wireguard*

Change-Id: I6d89922b1637239fbd4f9e2e04080fbb5a342a63
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3185559
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>