crash: Support LaCros build times.
When LaCros launches, the OS may be substantially more out-of-date than
the browser, as the two will update separately. So, use the new
"build_time_millis" meta field (added in https://crrev.com/c/3141938) to
determine the build time of the browser separately from the build time
of the OS, and use that to determine whether to send.

This change requires us to *stop* checking whether the OS build time is
too old before examining any reports, because we could have a too-old OS
but still a relatively new LaCros.  We'll continue to check each report
for a too-old OS time, just adding the check on the LaCros version as
well.

BUG=b:186665615
TEST=cros_workon_make --board=jacuzzi --test crash-reporter

Change-Id: I4d1c8271f17d4d59e7909696aed215c3bd95b27c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3142178
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>