shill: portal: retry detection on secondary networks
Until now, if shill portal detection fails to validate a network that is
not currently the highest priority network, additional validation
attempts are not scheduled.

This can negatively impact user experience for instance on devices with
a Cellular connection if validation happens to fail spuriously when the
device connects to WiFi: the device will not retry validating the WiFi
network and the device will keep using Cellular, incurring extra cost
for the user.

This patch changes this behavior such that retrying portal detection
detection and connectivity validation behaves the same way for the
highest priority network and all secondary networks.

Note #1: to force validation to fail on a given interface, it is
convenient to block the DNS replies with (IPv4 case):
  $ iptables -v -t filter -I INPUT -p udp --sport 53 -i <iface> -j DROP
This allows for instance to keep the ssh connection undisturbed.

Note #2: portal detection is effectively started through a single
function in Device::StartPortalDetection(). There are currently four
different codepaths that can lead to StartPortalDetection():

  a) Device::SetupConnection(), which happens on IP configuration
  changes. This is the main codepath that normally starts portal
  detection for a newly connected Service.

  b) Service::OnPropertyChanged -> Device::RestartPortalDetection().
  This codepath only happens if kCheckPortalProperty or
  kProxyConfigProperty of the currently selected Service has changed.

  c) Manager::RecheckPortal() -> Device::RequestPortalDetection(). This
  codepath implements the Manager DBus RecheckPortal() API.

  d) Manager::SortServices() -> Device::RequestPortalDetection(). This
  codepath is only triggered if a physical network currently already in
  a portal state becomes the new default network.

Since it is not necessary anymore to "restart" portal detection for a
secondary network that is not ONLINE and becomes the default network,
the last codepath d) is changed to restart portal detection from
scratch. This allows to reset the retry delay to the minimum value.

BUG=b:198817234
TEST=unit tests. Forced validation to fail on ethernet while cellular
was connected, observed shill retrying network validation on ethernet.
Removed iptables filter and observed shill validate ethernet and select
it as the default network.

Change-Id: I02e7ae557e797ea851890ad3565f853077ee9fab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3145035
Tested-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Garrick Evans <garrick@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>