dns-proxy: Fully disable DNS proxy when the device is not online
Previously, when the device is not yet online, DNS proxy
redirection rule will not be changed. On a special case of
captive portal, when the device is not online, name resolutions
might happen. This will fail captive portal auth flow because
the old name servers will be used instead.

In this patch, when the device is not yet online, the proxy will
be disabled by removing the redirection rule and reverting back
resolv.conf.

BUG=b:197256004
TEST=Connect to a captive portal with DNS proxy running;
     Observe that when the authentication is not done:
       - /etc/resolv.conf is not pointing to the proxy,
       - iptables redirection rules are removed.
     Observe that after the authentication is done:
       - /etc/resolv.conf is pointing to the proxy,
       - iptables redirection rules are properly configured.

Change-Id: I16a4b9603fac999308024bb4efa3df854d0ac967
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3143568
Tested-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Reviewed-by: Garrick Evans <garrick@chromium.org>
Commit-Queue: Jason Jeremy Iman <jasongustaman@chromium.org>