shill: Implementation of CQM listener
Kernel has capability to publish connection quality monitor
signals to the userspace over the netlink socket. In shill,
we can use these events to better identify issues in the
link quality such as Beacon losses, packet losses.

This implementation is to create listener for CQM notifications,
and to take actions in response to those events.
Specifically for Beacon and Packet losses, FW dump is triggered.
Between successive FW dumps there is a cool down period of 6 mins.
Additionally, metrics are published for Beacon and Packet losses.
The Beacon/Packet Loss notifications can come in successive intervals
during FW issue and if it goes beyond a threshold, the wifi driver
will initiate a disconnect. This acts as a governor to not continuously
bombard the userspace with excessive CQM notifications. Additional
details and design considerations on CQM is mentioned in the doc
go/cros-wifi-dynamic-data.

BUG=b:194945109
TEST=Verified by simulating fake beacon losses and packet losses.
     The fake beacon and packet losses where triggered by hacking
     the kernel driver to periodically publish the beacon and packet
     losses. The test kernel patch is crrev/c/3080639
TEST=Unit test are written to trigger CQM by replaying captured data
     from the Netlink socket.
TEST=FEATURES=test USE=asan emerge-dedede shill

Disallow-Recycled-Builds: test-failures
Change-Id: Ie503c0e8073755901f250663f8ae64a192dc13b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3080399
Commit-Queue: Abhishek Kumar <kuabhs@chromium.org>
Tested-by: Abhishek Kumar <kuabhs@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Joshua Emele <jemele@chromium.org>