debugd: Clarify --output-file option for packet capture
The error message on packet_capture_tool gives the usage of the command
when user supplies wrong options. Clarify in the command usage that the
output file input is only for internal calls from packet_capture_tool to
the helper executables and is not available for the runs of the command
from Crosh or other user-facing tools.

BUG=b:199170846
TEST=Run "packet_capture" on Crosh. The error message should have an
explanation about --output-file option.

Change-Id: Ied2ada86f1f89837d7b9a6f1a3030818860a5cd8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3147593
Tested-by: Irem Uğuz <iremuguz@google.com>
Commit-Queue: Irem Uğuz <iremuguz@google.com>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>