cryptohome: migrate callbacks in FingerprintManager
StartSessionCallback should be OnceCallback - it is used once or passed
to proxy_->StartAuthSessionAsync() which takes OnceCallback.

ResultCallback should be a RepeatingCallback - it may be called
multiple times from ProcessRetry(), which is called from
OnAuthScanDone(). The last method is used as the first argument to
default_proxy_->ConnectToAuthScanDoneSignal(), which takes a
RepeatingCallback.

BUG=chromium:1205291
TEST=USE=test emerge-betty cryptohome
TEST=cros_workon_make --board=betty chromeos-base/cryptohome --test

Change-Id: I0f82471909edb85c1a9df2528033c1cf40b3d922
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3147855
Tested-by: Tomasz Tylenda <ttylenda@chromium.org>
Reviewed-by: John L Chen <zuan@chromium.org>
Commit-Queue: Tomasz Tylenda <ttylenda@chromium.org>