vm_tools: Share ARCVM stub volumes using virtio-fs
We are going to use virtio-fs for ARCVM's stub volumes (MyFiles and
removable media).
They are shared in the same way as ARCVM's /data/media except that the
posix_acl option is set to false. This is because filesystems like FAT32
don't support xattrs. In order to do that, a new parameter posix_acl is
added to CreateSharedDataParam.

BUG=b:182870415
TEST=Mount "stub" on the Android side, and confirm that MyFiles and
TEST=removable media are available there.

Change-Id: Ie13462998e8bde0b0d212a81f6153788934f1f24
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3149796
Tested-by: Youkichi Hosoi <youkichihosoi@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Risan <risan@chromium.org>
Commit-Queue: Youkichi Hosoi <youkichihosoi@chromium.org>