cryptohome: [USS] Simplify field presence checks
As was confirmed by the Flatbuffers team, the IsFieldPresent checks can
be replaced with comparing the getter's result with null. The latter
form is slightly more concise, as it doesn't require using the VT_
constants.

Also improve a bit the error reporting in case some field is missing,
by mentioning the field name in the error log.

BUG=b:194953754
TEST=cros_workon_make --board=$BOARD --test cryptohome

Change-Id: Iec61d79d80e7ea3580062933c71adfb3fae2967c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3151842
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Greg Kerr <kerrnel@chromium.org>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>