typecd: Add Port pointer to Partner class
The Partner may require access to some information regarding the Port
(for determining what type of partner it itself is for metrics
reporting). So, add a pointer to the Port while constructing the
Partner.

Make it as a separate constructor so that existing unit tests don't all
need to be updated (since the Port pointer is only useful for metrics).

BUG=b:195056095
TEST=All unit tests pass.

Change-Id: I73fb22422b0a643e8bfef661e789cb82d5266dbf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3152409
Reviewed-by: Benson Leung <bleung@google.com>
Tested-by: Prashant Malani <pmalani@chromium.org>
Commit-Queue: Prashant Malani <pmalani@chromium.org>