cryptohome: Improve fuzzing via blob_mutator
Change the algorithm used in blob_mutator to increase the chance of the
fuzzing framework finding an interesting input.

This is achieved by switching from interpreting the fuzzer input as
enums + 32-bit integers (i.e., commands with arguments), and switch to
operating with masks that build the input byte-by-byte in a more direct
fashion. I.e., the backslash byte is now interpreted as "copy one byte
from the source blob", and any other byte is interpreted as "insert
this new byte". There are other combinations with backslashes - see the
comments in the code for the details.

This heuristic proved to be much more successful in quickly generating
interesting inputs, for example in generating structurally correct
mutated flatbuffers. In principle, this heuristic is based on the
existing implementation in
FuzzedDataProvider::ConsumeRandomLengthString().

BUG=b:199547188
TEST=USE=fuzzer emerge-$BOARD cryptohome

Change-Id: I6ec9b9797528e8f2ff3903dbad627b77527f71f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3151851
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>