vm_tools: load tools disk from arbitrary DLC
This a partial-reland of
commit fc00ca787dac5d2240082d655ce268ca5a1e8ae5. That was partially rolled back
in commit 0bc7ecfc1f6d97c16de306fa13036a2a6c656914 in order to restore
cros-termina component support, but was done in such a way that passing
a tools disk via DLC was broken on non-termina VMs.

This fixes that by:
* removing the non-termina/non-dlc early exit
* ensuring we only error out with "Termina component is not loaded" if
  we both provided an empty vm DLC ID and is_termina is set.

BUG=b:189120779,chromium:1228109
TEST=Manual build and deploy vm_host_tools, successfully start Bruschetta
and Crostini with DLC enabled, successfully start Crostini with DLC
disabled.

Change-Id: I909ce906a59656ccaf2cc8975696009a27cd6851
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3160430
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Reviewed-by: Fergus Dall <sidereal@google.com>
Tested-by: Mike Gerow <gerow@google.com>
Commit-Queue: Mike Gerow <gerow@google.com>