secanomalyd: Don't count W+X nsfs mounts.
Because of how the code was structured, we were not reporting the
anomaly event for nsfs mounts (bind-mounts of /proc/<pid>/ns/ files),
but we were adding these mounts to the dictionary tracking W+X mounts,
and therefore uploading them as part of the W+X mount count metric.

These mounts are already understood and do not represent an anomaly.
Reorder the code so that the nsfs check happens before adding the mount
to the dictionary.

A follow-up CL will report these mounts as a distinct anomaly event.
While these mounts are understood, we still want to get rid of them so
that we can simplify uploading logic.

BUG=chromium:1249652
TEST=New unit tests.
TEST=Deploy to device, run, see mounts reported in logs.
TEST=See nsfs mounts ignored in logs.
TEST=Open chrome:histograms, see:
TEST=WX mount count of 2: / and /usr/local.

Disallow-Recycled-Builds: test-failures
Change-Id: I9acd836f8dd249cb7bf34d0b7b65e6b3d472b1e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3160752
Tested-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Jorge Lucangeli Obes <jorgelo@chromium.org>
Reviewed-by: Nicole Anderson-Au <nvaa@google.com>