shill: vpn: Remove "rekey" and "IPsecTimeout" in new L2TPIPsecDriver
These two properties are only defined in the .cc file but not in
service-api.txt and dbus-constants.h, so they cannot be set by users
normally.

For the rekey property, there is no single property in swanctl.conf
mapping to it, but the behavior is controlled by several properties
instead. See
https://wiki.strongswan.org/projects/strongswan/wiki/ExpiryRekey

For the IPsecTimeout property, currently it is only used in
vpn-manager to control the timeout of an IPsec connection, but not used
in ipsec.conf.

This patch removes them for simplicity. We can add them back if we find
they are helpful for debugging later.

BUG=b:165170125
TEST=unit_tests

Change-Id: I9b427f32fd2bfcc7a67efbe6fed7d2ab1e297a58
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3161733
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>