shill: vpn: Set timeout when calling `swanctl --initiate`
`swanctl --initate` accepts a timeout option: if the option is set and
the connection is not established in the given time, swanctl will exit
with 1, and then IPsecConnection will trigger the connection failure
flow and stop charon in turn.

Also correct the timeout value used in the new L2TPIPsecDriver. The two
timeout values are the same as the ones used in the current
implementation correspondly.

BUG=b:165170125
TEST=Set up a L2TP/IPsec service with a nonexistent server, observed
  that service failed after 30 seconds, and verified that no charon
  process is running.

Change-Id: I20ecf4d353dffda4bd098fa8e0612ce359bc1615
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3161735
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>