shill: Group minijail options into a struct in ProcessManager
Currently we have several arguments related to minijail for
StartProcessInMinijail() and StartProcessInMinijailWithPipes() in
ProcessManager. The readability is not good since the argument list is
long, and it's hard to add a new argument for them since every call site
needs to be modified.

This patch group these arguments into a single struct, which will
increase the readability and easy to be extended.

BUG=b:165170125
TEST=tast run $DUT network.VPNConnect.*

Change-Id: I1ff1426a11e55dce52fe977b3e989b0c4da7b2cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3162356
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>