cryptohome: replace AuthBlockState proto with structs.
Test GetDoubleWrappedCompatAuthBlockState in
vault_keyset_unittest.cc is expanded to
two tests: Previously one GetAuthBlockState()
returns false for failure, which is not checked.
And the proto pointer still gets assigned
a value. We fix the logic with the new struct. It is
now properly checked, with the passed in struct
pointer correctly pointing to nullptr.
Then we add a test for success.

BUG=b:195641747
TEST=FEATURES=test emerge-betty-pi-arc cryptohome

Change-Id: I4b9ef98ca492623a1e76cf75510fd7ecba619829
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3114891
Tested-by: Zi Lin <lziest@chromium.org>
Commit-Queue: Zi Lin <lziest@chromium.org>
Reviewed-by: Greg Kerr <kerrnel@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>