cryptohome: Add support for using kernel keyring with dm-crypt
dm-crypt (from v1.15.0) supports the use of the kernel keyring to
provision keys from. This change adds support for using the keyring
if dm-crypt supports it.

The key is inserted into the thread-specific keyring as a "logon"
key - this ensures that key data is not accessible from userspace
once it has been inserted into the kernel. Additionally, using the
thread-specific keyring ensures that (1) no other users can attempt
to link to this key (2) in the unlikely event that cryptohome crashes
while setting up the key, the reference to this key is dropped and
the key is deleted automatically. Once the dm-crypt device has been
created, remove the key from the keyring.

BUG=b:198473496
TEST=Mount encrypted stateful partition with dm-crypt container:
1) dmsetup table shows keyring descriptor instead of raw keys.
2) /proc/keys no longer show the encryption key.

Change-Id: I849a45071edd584c12395c45a3b7a34d33bc5a2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3163208
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Commit-Queue: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Daniil Lunev <dlunev@chromium.org>